requirements.txt
pandas
pytz
sqlalchemy
pytest
u8darts[torch]
matplotlib
mariadb
python-dotenv
